+++ CHANGELOG.md
@@ -17,7 +17,21 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## [7.38.2] - 2024-04-23
+## [7.40.0] - 2024-04-30
+### Added
+- Datapoint Subscriptions now support status codes.
+
+## [7.39.0] - 2024-04-25
+### Added
+- Support for internally managed groups (inside CDF, as opposed to the external identity provider).
+
+## [7.38.3] - 2024-04-25
+### Improved
+- The classes `WorkflowUpsert`, `Filter`, `Query`, `Node`, `Edge`, `Container`, `Document`, and
+  `Transformation` which are used for parsing API responses were not handling adding new parameters in
+  the API correctly. These are now future-proofed.
+
+## [7.38.2] - 2024-04-24
 ### Added
 - Added new parameter `function_external_id` to `FunctionScheduleAPI.create` as a convenience to the user. Note
   that schedules must be attached to a Function by (internal) ID, so a lookup is first done on behalf of the user.